M document.json; M document/new-format/app/finder.txt; M document/new-format/app/terminal.txt; M document/new-format/file/note.txt; M document/new-format/other/buffers.txt; M document/new-format/other/other.txt; M document/new-format/visual/highlight.txt; M document/new-format/visual/other.txt; M other/quick.fish; M raw

+++ other/quick.fish
@@ -5,12 +5,15 @@ while true;
     python .merg.py
     cd ../other
     set b (python non-documented.py|head -1)
-    echo $b|xsel -b
-    git clone "https://www.github.com/$b" /tmp/gitfile --depth 1 -q
-    pushd .
-    cd /tmp/gitfile
-    bat (fd readme.\* -d 1) -p --paging=always
-    popd
-    rm -fr /tmp/gitfile
-    cd ..
+    echo $b|wl-copy
+    if not git clone "https://www.github.com/$b" /tmp/gitfile --depth 1 -q
+        read
+    else
+        pushd .
+        cd /tmp/gitfile
+        bat (fd readme.\* -d 1) -p --paging=always
+        popd
+        rm -fr /tmp/gitfile
+        cd ..
+    end
 end
